21-06-24 학생 성적관리 완료 / 학적과 학생등록 tiles 적용 / 스캐줄러 컨트롤러 분리

+++ shms/src/main/java/kr/ac/shms/lms/student/vo/AttendVO.java
@@ -68,7 +68,7 @@ public class AttendVO {
 	private Integer jg_cnt;
 	private Integer jt_cnt;
 	private Integer gs_cnt;
-	private Integer attend_per;
+	private Double attend_per;
 	
 	private String lec_term;
 	private String dayotw_nm;
